Integrates FNV-1a stable device index and net-snmp iterator bug fix into
master, resolving conflicts against state-based sensor traps and related
features added since the branch point.

Key fix: net-snmp's table iterator silently drops GETNEXT results for OID
sub-identifiers >= 2^31.  FNV hash is now capped to 31 bits (& 0x7FFFFFFF)
and devices are sorted by index after initial scan.  Without this cap,
drives whose FNV(serial|model) hash >= 2^31 are invisible in all SNMP walks.

Also included from patch-split:
- Move model/serial/firmware/WWN to common device table (cols 10-14)
- consec_fail_count hysteresis for poll-failure trap
- estimated_completion, error_timestamp fields on self-test/error rows
- SATA change tracking subtree (.4.1.2)

+++ src/snmp_common_mib.cpp
@@ -35,9 +35,8 @@ device_get_next(void **loop_ctx, void **data_ctx,
     CacheDeviceRow &row = g_cache.devices[idx];
     *loop_ctx = (void*)(uintptr_t)(idx + 1);
     *data_ctx = &row;
-    u_long uval = (u_long)row.index;
-    snmp_set_var_typed_value(put_idx, ASN_UNSIGNED,
-                             (u_char*)&uval, sizeof(uval));
+    uint32_t uval = row.index;
+    snmp_set_var_typed_value(put_idx, ASN_UNSIGNED, (u_char*)&uval, sizeof(uval));
     return put_idx;
 }
